Fix BCELoss crash under mixed precision
nn.BCELoss operates on post-sigmoid probabilities, which PyTorch's
autocast explicitly disallows computing in fp16 (numerically unstable
near 0/1, and it raises rather than silently degrading) — this crashed
the CUDA training run as soon as --device cuda hit the first loss
computation. The forward pass (conv/pooling, the expensive part) still
runs under autocast for the fp16 speedup; the loss itself is now
computed in fp32 outside that context, which is the standard fix for
this exact restriction.

Note: this sandbox has no GPU, so this is fixed against the traceback
and PyTorch's documented autocast op restrictions, not verified by
actually exercising the CUDA path here — needs confirmation on real
hardware.

+++ tennis_tracker/tracknet.py
@@ -353,7 +353,10 @@ def train(
             optimizer.zero_grad()
             with torch.amp.autocast(device_type, enabled=amp_enabled):
                 predictions = model(frames)
-                loss = loss_fn(predictions, targets)
+            # BCELoss on post-sigmoid probabilities is numerically unstable
+            # (and disallowed by autocast) in fp16 — the heavy conv/pooling
+            # work above still runs in fp16, but the loss itself needs fp32.
+            loss = loss_fn(predictions.float(), targets)
             scaler.scale(loss).backward()
             scaler.step(optimizer)
             scaler.update()
